requirements.txt
configparser==5.0.2
pkg-resources==0.0.0
psycopg2==2.8.6
pyserial==3.5
pytz==2021.1
tzlocal==2.1
